// apps/web/src/generated/api/models/nestCreateDtoJoinPolicy.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * ThreadNest API
 * OpenAPI spec version: 1.0.0
 */

export type NestCreateDtoJoinPolicy = typeof NestCreateDtoJoinPolicy[keyof typeof NestCreateDtoJoinPolicy]

export const NestCreateDtoJoinPolicy = {
  OPEN: 'OPEN',
  BY_REQUEST: 'BY_REQUEST',
  BY_INVITE: 'BY_INVITE'
} as const